move monthly topic to page folder
